server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
diskcache==5.6.3
dotenv==0.9.9
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
setuptools==78.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
python-jose[cryptography]==3.5.0
pyjwt==2.10.1